Changed Websocket controller, worked on data validation

+++ src/main/java/com/takehometask/BetMachine/model/BetResponseModel.java
@@ -1,6 +1,8 @@
 package com.takehometask.BetMachine.model;
 
+import java.math.BigDecimal;
+
 public record BetResponseModel(
-        float win
+        BigDecimal win
 ) {
 }
